Actually fix hgss textbox type mismatch

+++ include/graphics.h
@@ -6,8 +6,8 @@ extern const u8 gSignpostWindow_Gfx[];
 extern const u8 gMessageBox_Gfx[];
 extern const u16 gMessageBox_Pal[];
 
-extern const u32 gSignpostWindow_HGSS_Gfx[];
-extern const u32 gMessageBox_HGSS_Gfx[];
+extern const u8 gSignpostWindow_HGSS_Gfx[];
+extern const u8 gMessageBox_HGSS_Gfx[];
 extern const u16 gMessageBox_HGSS_Pal[];
 
 // pokeballs
